refactor: move IdentitySetupInitializer to new package
