+++ zeebe/engine/src/main/java/io/camunda/zeebe/engine/processing/identity/initialize/IdentitySetupInitializer.java
@@ -5,7 +5,7 @@
  * Licensed under the Camunda License 1.0. You may not use this file
  * except in compliance with the Camunda License 1.0.
  */
-package io.camunda.zeebe.engine.processing.user;
+package io.camunda.zeebe.engine.processing.identity.initialize;
 
 import static io.camunda.zeebe.protocol.record.value.AuthorizationScope.WILDCARD;
